+++ DisplayManager.h
@@ -20,6 +20,7 @@ public:
 
     void begin();
     void handle();
+    void showInfoMessage(const char* line1, const char* line2);
 
 private: